Add users crud

+++ app/Actions/Jetstream/DeleteUser.php
@@ -2,6 +2,7 @@
 
 namespace App\Actions\Jetstream;
 
+use App\Models\User;
 use Laravel\Jetstream\Contracts\DeletesUsers;
 
 class DeleteUser implements DeletesUsers
@@ -14,8 +15,10 @@ class DeleteUser implements DeletesUsers
      */
     public function delete($user)
     {
+        abort_if($user->isAdministrator() && User::admin()->count() <= 1, 403, 'Must have at least one administrator account left.');
+
         $user->deleteProfilePhoto();
-        $user->tokens->each->delete();
+
         $user->delete();
     }
 }
